Add Strategies
Reason: Case -Sensitive and -Insensitive deserve their own header
