--- ibragimov.imil/T2/label.cpp
@@ -1,10 +1,0 @@
-#include "label.hpp"
-
-bool ibragimov::formatters::CaseSensitive::isSame(const char lChar, const char rChar)
-{
-  return (lChar == rChar);
-}
-bool ibragimov::formatters::CaseInsensitive::isSame(const char lChar, const char rChar)
-{
-  return (std::tolower(lChar) == std::tolower(rChar));
-}
