+++ ibragimov.imil/T2/strategies.hpp
@@ -1,0 +1,19 @@
+#ifndef STRATEGIES_HPP
+#define STRATEGIES_HPP
+
+namespace ibragimov
+{
+  namespace strategies
+  {
+    struct CaseSensitive
+    {
+      static bool isSame(const char, const char);
+    };
+    struct CaseInsensitive
+    {
+      static bool isSame(const char, const char);
+    };
+  }
+}
+
+#endif
